+++ infra/binder/src/main/java/org/apache/shardingsphere/infra/binder/engine/type/DDLStatementBindEngine.java
@@ -24,6 +24,7 @@ import org.apache.shardingsphere.infra.binder.engine.statement.ddl.CreateIndexSt
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.CreateTableStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.CursorStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.ddl.DropTableStatementBinder;
+import org.apache.shardingsphere.infra.binder.engine.statement.ddl.RenameTableStatementBinder;
 import org.apache.shardingsphere.infra.hint.HintValueContext;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.AlterTableStatement;
@@ -32,6 +33,7 @@ import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.CreateT
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.CursorStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.DDLStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.DropTableStatement;
+import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.RenameTableStatement;
 
 /**
  * DDL statement bind engine.
@@ -64,6 +66,9 @@ public final class DDLStatementBindEngine {
         }
         if (statement instanceof DropTableStatement) {
             return new DropTableStatementBinder().bind((DropTableStatement) statement, binderContext);
+        }
+        if (statement instanceof RenameTableStatement) {
+            return new RenameTableStatementBinder().bind((RenameTableStatement) statement, binderContext);
         }
         if (statement instanceof CreateIndexStatement) {
             return new CreateIndexStatementBinder().bind((CreateIndexStatement) statement, binderContext);